found bug ins motion 2

+++ object.h
@@ -1,20 +1,20 @@
 class obj
 {
 public:
-obj(float a,float b,float c);
-~obj();
-
-float t;
-
-float x0,y0,z0;
-float ax0,ay0,vx0,vy0;
-
-float x,y,z;
-float ax,ay,vx,vy;
-
-void setPosition(float a,float b,float c);
-void set_a(float x,float y);
-void set_v(float x,float y);
-void reset();
-void getValues(float t,int s,float x1,float y1);
+	obj(double a,double b,double c);
+	~obj();
+	
+	double t;
+	
+	double x0,y0,z0;
+	double ax0,ay0,vx0,vy0;
+	
+	double x,y,z;
+	double ax,ay,vx,vy;
+	
+	void setPosition(double a,double b,double c);
+	void set_a(double x,double y);
+	void set_v(double x,double y);
+	void reset();
+	void getValues(double t,int s,double x1,double y1);
 };
